pyproject.toml
[project]
name = "paper-paper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "networkx>=3.5",
    "paper-network>=0.0.1",
    "scipy>=1.16.2",
]
